CMS-124 Static analysis for Prestashop (php 8.1)

See merge request stancer/ecommerce-module/prestashop!34

+++ classes/StancerApiConfig.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * Stancer PrestaShop
  *
@@ -26,7 +27,7 @@ class StancerApiConfig
     /** @var int|null API Timeout */
     public $timeout;
 
-    /** @var string Auth limit */
+    /** @var string|null Auth limit */
     public $authLimit;
 
     /** @var bool API is configured ? */
@@ -41,7 +42,7 @@ class StancerApiConfig
     {
         $this->mode = Configuration::get('STANCER_API_MODE') ?: Stancer\Config::TEST_MODE;
         $this->host = Configuration::get('STANCER_API_HOST');
-        $this->timeout = Configuration::get('STANCER_API_TIMEOUT');
+        $this->timeout = (int) Configuration::get('STANCER_API_TIMEOUT');
         $this->authLimit = Configuration::get('STANCER_AUTH_LIMIT');
         $this->isConfigured = $this->isConfigured();
     }
